requirements.txt
django
djangorestframework
pywinauto
pyperclip
pyautogui